+++ x-pack/platform/test/cases_api_integration/security_and_spaces/tests/common/attachments_framework/validation.ts
@@ -9,6 +9,7 @@ import type {
   UnifiedReferenceAttachmentPayload,
   UnifiedValueAttachmentPayload,
 } from '@kbn/cases-plugin/common/types/domain';
+import type { AttachmentRequest } from '@kbn/cases-plugin/common/types/api';
 import type { FtrProviderContext } from '../../../../common/ftr_provider_context';
 import { postCaseReq } from '../../../../common/lib/mock';
 import { createCase, deleteAllCaseItems, bulkCreateAttachments } from '../../../../common/lib/api';
@@ -71,6 +72,49 @@ export default ({ getService }: FtrProviderContext): void => {
               data: { content: 'invalid' },
               owner: 'securitySolutionFixture',
             },
+          ],
+          expectedHttpCode: 400,
+        });
+      });
+    });
+
+    // Legacy `externalReference` / `persistableState` wire shapes are only accepted
+    // for migrated type ids (EXTERNAL_REFERENCE_TYPE_MAP / PERSISTABLE_STATE_TYPE_MAP),
+    // which route through the transformer to a registered unified type. An unmapped id
+    // has no unified target and must be rejected — this replaces the removed
+    // custom-ER/PS-type registration coverage.
+    describe('legacy wire shapes with an unmapped type', () => {
+      it('400s for a legacy externalReference with an unmapped type id', async () => {
+        const postedCase = await createCase(supertest, postCaseReq);
+        await bulkCreateAttachments({
+          supertest,
+          caseId: postedCase.id,
+          params: [
+            {
+              type: 'externalReference',
+              externalReferenceId: 'ref-id',
+              externalReferenceStorage: { type: 'elasticSearchDoc' },
+              externalReferenceAttachmentTypeId: 'not.mapped.external.reference',
+              externalReferenceMetadata: null,
+              owner: 'securitySolutionFixture',
+            } as unknown as AttachmentRequest,
+          ],
+          expectedHttpCode: 400,
+        });
+      });
+
+      it('400s for a legacy persistableState with an unmapped type id', async () => {
+        const postedCase = await createCase(supertest, postCaseReq);
+        await bulkCreateAttachments({
+          supertest,
+          caseId: postedCase.id,
+          params: [
+            {
+              type: 'persistableState',
+              persistableStateAttachmentTypeId: 'not.mapped.persistable.state',
+              persistableStateAttachmentState: { foo: 'foo' },
+              owner: 'securitySolutionFixture',
+            } as unknown as AttachmentRequest,
           ],
           expectedHttpCode: 400,
         });
