+++ x-pack/platform/test/cases_api_integration/security_and_spaces/tests/common/attachments_framework_legacy/legacy_route_compat.ts
@@ -6,12 +6,15 @@
  */
 
 import type http from 'http';
+import expect from '@kbn/expect';
 import {
   CASE_ATTACHMENT_SAVED_OBJECT,
   DASHBOARD_ATTACHMENT_TYPE,
   DASHBOARD_SO_TYPE,
+  SECURITY_TIMELINE_ATTACHMENT_TYPE,
 } from '@kbn/cases-plugin/common/constants';
 import { ALERTING_CASES_SAVED_OBJECT_INDEX } from '@kbn/core-saved-objects-server/src/saved_objects_index_pattern';
+import type { AttachmentRequestV2 } from '@kbn/cases-plugin/common/types/api';
 import type { Client } from '@elastic/elasticsearch';
 import { ObjectRemover as ActionsRemover } from '../../../../../alerting_api_integration/common/lib';
 import type { FtrProviderContext } from '../../../../common/ftr_provider_context';
@@ -19,6 +22,7 @@ import { postCaseReq } from '../../../../common/lib/mock';
 import {
   createCase,
   createCaseWithConnector,
+  createComment,
   deleteAllCaseItems,
   getComment,
   getServiceNowSimulationServer,
@@ -114,6 +118,27 @@ export default ({ getService }: FtrProviderContext): void => {
         connectorId: connector.id,
         expectedHttpCode: 200,
       });
+    });
+
+    // `security.timeline` is a unified-only type with no legacy representation, so it cannot
+    // be persisted as a legacy `cases-comments` SO. With the flag OFF the write path rejects
+    // it with an actionable 400 instead of an opaque 500.
+    it('400s when creating a unified-only security.timeline attachment', async () => {
+      const postedCase = await createCase(supertest, postCaseReq);
+      const response = (await createComment({
+        supertest,
+        caseId: postedCase.id,
+        params: {
+          type: SECURITY_TIMELINE_ATTACHMENT_TYPE,
+          owner: 'securitySolutionFixture',
+          attachmentId: 'timeline-1',
+          metadata: { title: 'My timeline' },
+        } as unknown as AttachmentRequestV2,
+        expectedHttpCode: 400,
+      })) as unknown as { statusCode: number; message: string };
+
+      expect(response.statusCode).to.be(400);
+      expect(response.message).to.contain('has no legacy representation');
     });
   });
 };
