--- x-pack/platform/test/functional_with_es_ssl/plugins/cases/server/attachments/persistable_state.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import type { PersistableStateAttachmentTypeSetup } from '@kbn/cases-plugin/server/attachment_framework/types';
-
-export const getPersistableStateAttachmentServer = (): PersistableStateAttachmentTypeSetup => ({
-  id: '.test',
-});
